+++ Dyalect/Compiler/CompilerError.cs
@@ -82,6 +82,12 @@
 
         CtorNotIterator = 239,
 
-        PositionalArgumentAfterKeyword = 240
+        PositionalArgumentAfterKeyword = 240,
+
+        CtorOnlyStatic = 241,
+
+        CtorOnlyPascal = 242,
+
+        // = 243
     }
 }