+++ Dyalect/Compiler/OpStackHelper.cs
@@ -85,13 +85,11 @@
             0,  //TypeCheckT
             0,  //Start
             0,  //End
-            0,  //NewType
+            1,  //NewType
             1,  //TypeS
             1,  //TypeST
             0,  //CtorCheck
             1,  //Unbox
-            0, //Rethrow
-            0, //PopErr
             0, //IsNull
         };
     }