+++ Dyalect/Parser/Model/DFunctionDeclaration.cs
@@ -13,9 +13,14 @@ namespace Dyalect.Parser.Model
 
         public string Name { get; set; }
 
-        internal bool IsStatic => Attribute is FunctionAttribute.Static or FunctionAttribute.Constructor;
+        internal bool IsStatic { get; set; }
 
-        public FunctionAttribute Attribute { get; set; }
+        private bool? isConstructor;
+        internal bool IsConstructor
+        {
+            get => isConstructor ?? IsMemberFunction && Name is not null && char.IsUpper(Name[0]);
+            set => isConstructor = value;
+        }
 
         public bool IsIterator { get; set; }
 
@@ -45,6 +50,9 @@ namespace Dyalect.Parser.Model
                 return;
             }
 
+            if (IsStatic)
+                sb.Append("static ");
+
             if (Name is not null)
                 sb.Append("func ");
 
@@ -65,25 +73,10 @@ namespace Dyalect.Parser.Model
             if (Name is not null || Parameters.Count > 1)
                 sb.Append(") ");
 
-            if (Attribute == FunctionAttribute.Constructor)
-                sb.Append("cons ");
-            else if (Attribute == FunctionAttribute.Static)
-                sb.Append("static ");
-            else if (Attribute == FunctionAttribute.Deconstructor)
-                sb.Append("decons ");
-
             if (Name is null)
                 sb.Append(" => ");
 
             Body.ToString(sb);
         }
-    }
-
-    public enum FunctionAttribute
-    {
-        None,
-        Static,
-        Constructor,
-        Deconstructor
     }
 }